Destructure Fact Sources for CMS
Due to the difference in structure of data coming through graphQL and data accessed using
entry.getIn on CMS, the only way to have a component that will access and render both corretly
is either have many parameters and pass subfields name and url of source
to the component or destructure the data on CMS. I chose the latter. The
difference in structure is also why the dialog inside LargeFactCard doesn't access the source
name and url.