Move hook inside React Context & wrap around app
I moved the hook that was determining the grid & list view out of facts
page and put it inside a react context in view-provider.js. The context
provider providers the state & function to change state. The context
consumer allows facts page to use the state. Gatsby-browser allows us to
use wrapRootElement to wrap the app inside the context so every page
has the context which includes the state controlling the view.