Add a list of contributors to about page
The gatsby-source-github-api doesn't support retrieving contributors
yet. The closest thing it can give is a list of collaborators which are different, so I used a different plugin.
I also filter dependabot out in the GraphQL query. I initally had a
regex query that fitlered out any login with
"bot" in it, but that could theoretically catch future contributors
whose login contain the words 'bot'.